+++ server/store/types/errors.go
@@ -1,4 +1,4 @@
-// Copyright 2021 Woodpecker Authors
+// Copyright 2022 Woodpecker Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,19 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package datastore
+package types
 
-import (
-	"github.com/woodpecker-ci/woodpecker/server/store/types"
-)
+import "database/sql"
 
-// wrapGet return error if err not nil or if requested entry do not exist
-func wrapGet(exist bool, err error) error {
-	if err != nil {
-		return err
-	}
-	if !exist {
-		return types.RecordNotExist
-	}
-	return nil
-}
+var RecordNotExist = sql.ErrNoRows
